Patch from Buğra Gedik - bgedik at gmail dot com
"I run into a problem. It cannot locate the file names and numbers for calls to functions that are in linked libraries, rather than in the executable.

The reason is that, the linked libraries may be located at arbitrary places within the virtual address space of the process. One needs to find where they are located and fix the addresses accordingly, before resolving the symbol via libbfd. This can be done by reading the /proc/<pid>/maps file, which lists where the loaded libraries are located within the address space."